requirements.txt
torch==2.3.1
open_clip_torch==2.32.0
opencv-python==4.11.0.86
pillow==11.1.0
tqdm==4.65.2
supervision==0.25.1
pandas